requirements.txt
rich>=13.7.1
psutil>=5.9.8
argon2-cffi>=23.1.0
